Integrate the core repoAuthorList to suggest cmd
Convert suggest to typescript.
