+++ packages/git-mob/src/git-commands.js
@@ -129,15 +129,6 @@ function topLevelDirectory() {
   return silentRun('git rev-parse --show-toplevel').stdout.trim();
 }
 
-/**
- * Returns a list of the existing authors for the git repository
- * including their names and email addresses
- * @returns {string} of output from git command
- */
-function shortLogAuthorSummary() {
-  return silentRun('git shortlog --summary --email --number HEAD').stdout.trim();
-}
-
 function getTemplatePath() {
   return get('commit.template');
 }
@@ -181,7 +172,4 @@ export const revParse = {
   gitPath,
   insideWorkTree,
   topLevelDirectory,
-};
-export const authors = {
-  shortLogAuthorSummary,
 };
